some small refactor signin and profile edit fetches
