+++ src/pages/Profile/helpers/updateUser.js
@@ -1,12 +1,14 @@
 import { useSelector } from 'react-redux';
 import useRequest from '../../../hooks/useFetch';
-import updateUser from '../../../services/profile/updateUser';
+import { getUserUpdateRequestData } from '../../../services/users';
 
 const useUpdateUserHook = () => {
   const makeRequest = useRequest();
   const token = useSelector((state) => state.signin.token);
   const update = async (edited) => {
-    await makeRequest(updateUser(token, edited._id, edited));
+    await makeRequest(
+      getUserUpdateRequestData({ token, id: edited._id, body: edited })
+    );
   };
   return update;
 };
