+++ src/services/users/index.js
@@ -10,6 +10,16 @@ export const getUserInvitationRequestData = ({ token, body }) =>
     body: JSON.stringify(body),
   });
 
+export const getUserUpdateRequestData = ({ token, id, body }) =>
+  new Request(`${MAIN_URL}users/${id}`, {
+    method: 'PUT',
+    headers: {
+      'Content-Type': 'application/json;charset=utf-8',
+      Authorization: `Bearer ${token}`,
+    },
+    body: JSON.stringify(body),
+  });
+
 export const getLimitedUsersRequestData = (token, limit, page, isAdmin) =>
   new Request(
     `${MAIN_URL}users${isAdmin ? `/all?limit=${limit}&page=${page}` : ''}`,
